refactor(Profit): 수익 관련 클래스 수정
수익은 정수 자료형 int 의 최댓값을 넘을 수 있으므로, 생성자를 더 큰 값까지 표현 가능한 long으로 변경

+++ src/main/java/domain/Profit.java
@@ -5,11 +5,11 @@ public class Profit {
     private final String percentageProfit;
     private final float PERCENTAGE = 100F;
 
-    public Profit(int money, int prizeMoney) {
+    public Profit(int money, long prizeMoney) {
         percentageProfit = calculatePercentageProfit(money, prizeMoney);
     }
 
-    private String calculatePercentageProfit(int money, int prizeMoney) {
+    private String calculatePercentageProfit(int money, long prizeMoney) {
         float profit = (Float.valueOf(prizeMoney) / Float.valueOf(money)) * PERCENTAGE;
         return String.format("%.1f", profit);
     }
